only start container if not started, stats update

+++ start-kaggle-container.sh
@@ -13,5 +13,10 @@ if [ ! -f $dest_data_file ] ; then
     curl --create-dirs $remote_source_file --output $dest_data_file
 fi
 
-# Mount the current working directory, hopefully including this git repo in its' directory structure
-docker run --rm -it -v $PWD:/home/jupyter -p 8080:8080 --env NOTEBOOK_DISABLE_ROOT=true gcr.io/kaggle-images/python /run_jupyter.sh 
+# start the container only if it is not already running
+running_images=`docker ps -a | grep kaggle-images`
+if [ -z "$running_images" ] ; then
+    docker ps
+    # Mount the current working directory, hopefully including this git repo in its' directory structure
+    docker run --rm -it -v $PWD:/home/jupyter -p 8080:8080 --env NOTEBOOK_DISABLE_ROOT=true gcr.io/kaggle-images/python /run_jupyter.sh 
+fi
